fix: alterando todas as migrações pra segurança

+++ db/migrate/20250404032753_create_i_produtos.rb
@@ -2,26 +2,25 @@
 
 class CreateIProdutos < ActiveRecord::Migration[7.2]
   def up
-    create_table :i_produtos do |t|
-      t.string     :nome
-      t.string     :slug
-      t.text       :descricao
-      t.decimal    :preco, precision: 10, scale:2
-      t.references :g_categoria, null: false, foreign_key: true
-      t.string     :arquivo_url
-      t.string     :imagem_url
-      t.string     :status
-      t.string     :created_by
-      t.string     :updated_by
-      t.datetime :deleted_at
-      t.timestamps
+    unless table_exists?(:i_produtos)
+      create_table :i_produtos do |t|
+        t.string     :nome,     null: false
+        t.string     :slug,     null: false
+        t.text       :descricao
+        t.decimal    :preco, precision: 10, scale: 2
+        t.references :g_categoria, null: false, foreign_key: true
+        t.string     :arquivo_url
+        t.string     :imagem_url
+        t.integer    :status, default: 1, null: false  
+        t.datetime   :deleted_at
+        t.timestamps
+      end
 
+      add_index :i_produtos, :slug, unique: true
     end
-    
-    add_index :i_produtos, :slug, unique: true
   end
 
   def down
-    drop_table :i_produtos
+    drop_table :i_produtos if table_exists?(:i_produtos)
   end
 end
