+++ db/migrate/20250415040351_create_i_promocao_produtos.rb
@@ -2,18 +2,20 @@
 
 class CreateIPromocaoProdutos < ActiveRecord::Migration[7.2]
   def up
-    create_table :i_promocao_produtos do |t|
-      t.references :i_promocao, foreign_key: true
-      t.references :i_produto,  foreign_key: true
-      t.decimal    :preco_promocional, precision: 10, scale: 2
-      t.string     :created_by
-      t.string     :updated_by
-      t.datetime   :deleted_at
-      t.timestamps
+    unless table_exists?(:i_promocao_produtos)
+      create_table :i_promocao_produtos do |t|
+        t.references :i_promocao, foreign_key: true
+        t.references :i_produto,  foreign_key: true
+        t.decimal    :preco_promocional, precision: 10, scale: 2
+        t.datetime   :deleted_at
+        t.timestamps
+      end
+
+      add_index :i_promocao_produtos, [:i_promocao_id, :i_produto_id], unique: true, name: 'index_promocao_produto_unique'
     end
   end
 
   def down
-    drop_table :i_promocao_produtos
+    drop_table :i_promocao_produtos if table_exists?(:i_promocao_produtos)
   end
 end
