+++ db/migrate/20250417150309_create_i_carrinhos.rb
@@ -2,19 +2,19 @@
 
 class CreateICarrinhos < ActiveRecord::Migration[7.2]
   def up
-    create_table :i_carrinhos do |t|
-      t.references :usuario, null: false, foreign_key: { to_table: :users }
-      t.string     :status
-      t.datetime   :criado_em
-      t.datetime   :atualizado_em
-      t.string     :created_by
-      t.string     :updated_by
-      t.datetime   :deleted_at
-      t.timestamps
+    unless table_exists?(:i_carrinhos)
+      create_table :i_carrinhos do |t|
+        t.references :usuario, null: false, foreign_key: { to_table: :users }
+        t.string     :status
+        t.datetime   :criado_em
+        t.datetime   :atualizado_em
+        t.datetime   :deleted_at
+        t.timestamps
+      end
     end
   end
 
   def down
-    drop_table :i_carrinhos
+    drop_table :i_carrinhos if table_exists?(:i_carrinhos)
   end
 end
